Improve controls Reduce memory - small enough for an Colecovision Port

+++ libs/console.c
@@ -49,7 +49,7 @@ void print_str(unsigned char x, unsigned char y, char *str, short offset) {
     }
 }
 
-void print_num(unsigned char x, unsigned char y, long num, short offset) {
+void print_num(unsigned char x, unsigned char y, unsigned int num, short offset) {
     char buffer[10+1]; 
     char *str;
